+++ src/Segment/Segment.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Smeghead\TextLinkEncoder\Segment;
 
-use Smeghead\TextLinkEncoder\Config\Settings;
+use Smeghead\TextLinkEncoder\Config\TextLinkEncoderSettings;
 
 interface Segment
 {
     public static function getSearchRegex(): string;
-    public function __construct(Settings $settings, string $segment);
+    public function __construct(TextLinkEncoderSettings $settings, string $segment);
     public function toHtml(): string;
 }