+++ src/Segment/TextSegment.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Smeghead\TextLinkEncoder\Segment;
 
-use Smeghead\TextLinkEncoder\Config\Settings;
+use Smeghead\TextLinkEncoder\Config\TextLinkEncoderSettings;
 
 final class TextSegment implements Segment
 {
@@ -14,7 +14,7 @@ final class TextSegment implements Segment
     }
 
     // @phpstan-ignore-next-line
-    public function __construct(private Settings $settings, private string $segment)
+    public function __construct(private TextLinkEncoderSettings $settings, private string $segment)
     {
     }